Remove ENABLE_NVIDIA toggle, build test separately
ENABLE_NVIDIA caused problems because the header file and
compiled code could differ if the flag used for compilation is not the
same as the one provided by the client using the hhal.

For example not specifying ENABLE_NVIDIA as a client
when the hhal was compiled with ENABLE_NVIDIA would make the
client think that the nvidia_manager had a different set of
internal properties, so not enough memory is allocated for it.